+++ lib/riot+compiler.js
@@ -1,3 +1,15 @@
-import riot from './browser/compiler/compile'
-export default riot
+import * as riot from './browser/index'
+import { compile, parsers } from './browser/compiler/index'
+import { extend } from './browser/util'
 
+function mount(a, b, c) {
+  var ret
+  compile(function () { ret = riot.mount(a, b, c) })
+  return ret
+}
+
+export default extend({}, riot, {
+  mount,
+  compile,
+  parsers
+})
