+++ package.json
@@ -59,10 +59,10 @@
     "chai": "^3.5.0",
     "cheerio": "^0.20.0",
     "coveralls": "^2.11.9",
-    "eslint": "^2.10.2",
+    "eslint": "^2.11.0",
     "glob": "^7.0.3",
     "istanbul": "^0.4.3",
-    "jsdom": "^9.2.0",
+    "jsdom": "^9.2.1",
     "karma": "^0.13.22",
     "karma-chrome-launcher": "^1.0.1",
     "karma-coverage": "^1.0.0",
