+++ test/specs/browser/core.spec.js
@@ -9,7 +9,6 @@ import {
   makeTag,
   defineTag
 } from '../../helpers/index'
-import riot from 'riot'
 
 // include special tags to test specific features
 import '../../tag/v-dom-1.tag'
