+++ test/specs/browser/if.spec.js
@@ -8,7 +8,6 @@ import {
   defineTag
 } from '../../helpers/index'
 
-import riot from 'riot'
 
 // include special tags to test specific features
 import '../../tag/if-mount.tag'
