+++ test/specs/browser/mixin.spec.js
@@ -1,5 +1,4 @@
 import { injectHTML } from '../../helpers/index'
-import riot from 'riot'
 
 const expect = chai.expect
