+++ test/specs/browser/style.spec.js
@@ -6,7 +6,6 @@ import {
   getRiotStyles
 } from '../../helpers/index'
 
-import riot from 'riot'
 
 // include special tags to test specific features
 import '../../tag/style-tag.tag'
